vulnerability show works!
refactored timeline to use ES6 classes
