+++ app/assets/javascripts/vulnerabilities/show.js
@@ -1,5 +1,5 @@
 import $ from 'jquery'
-import timelines from './timelines'
+import Timeline from './timeline'
 import SimpleDate from '../global/simpleDate'
 import {vhpMarkdown} from '../global/vhpMarkdown'
 
@@ -18,7 +18,7 @@ function initHTimeline(defaultZoom = 'vcc_to_fix'){
     defaultZoomValue: defaultZoom,
   }
 
-  var h = new timelines(settings);
+  var h = new Timeline(settings);
   h.init();
   return h;
 }
@@ -29,10 +29,10 @@ function initHTimeline(defaultZoom = 'vcc_to_fix'){
  * @returns {*} - array of sorted event hashes
  */
 function sortEvents(events) {
-    events.sort(function(a, b) {
-        return b.date.getTime() - a.date.getTime();
-    });
-    return events;
+  events.sort(function(a, b) {
+      return b.date.getTime() - a.date.getTime();
+  });
+  return events;
 }
 
 function populateHTimeline(hTimeline, vulnEvents) {
@@ -263,7 +263,7 @@ function load_succeded(vulnEvents, releases) {
   populateVTimeline(sortedEvents);
 
   $('#htimeline-loading').hide();
-  populateHTimeline(hTimeline, sortedEvents);
+  populateHTimeline(window.hTimeline, sortedEvents);
 }
 
 function load_failed(vulnEvents, releases) {
@@ -274,9 +274,8 @@ function load_failed(vulnEvents, releases) {
   $('#vtimeline-loading-failed').show();
 }
 
-
 export default function onVulnerabilitiesShow(){
-  var hTimeline = initHTimeline();
+  window.hTimeline = initHTimeline();
   window.events = []; // global to the page for resizing
   $.when(
     $.ajax(`/api/vulnerabilities/${vulnerability_id}/events`),
