+++ config/webpack/webpack.config.js
@@ -2,10 +2,11 @@ const path    = require("path")
 const webpack = require("webpack")
 
 // Extracts CSS into .css file
-const MiniCssExtractPlugin = require('mini-css-extract-plugin');
-// Removes exported JavaScript files from CSS-only entries
-// in this example, entry.custom will create a corresponding empty custom.js file
-const FixStyleOnlyEntriesPlugin = require('webpack-fix-style-only-entries');
+// not using css in webpack - using rails asset pipeline instead
+// const MiniCssExtractPlugin = require('mini-css-extract-plugin');
+
+// Analyze the size of our bundle
+const BundleAnalyzerPlugin = require('webpack-bundle-analyzer').BundleAnalyzerPlugin;
 
 module.exports = {
   mode: "production",
@@ -15,30 +16,35 @@ module.exports = {
   },
   module: {
     rules: [
-      { // Add CSS/SASS/SCSS rule with loaders
-        test: /\.s[ac]ss$/i,
-        use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
-      },
+      // not using css in webpack - using rails asset pipeline instead
+      // { // Add CSS/SASS/SCSS rule with loaders
+      //   test: /\.s[ac]ss$/i,
+      //   use: [MiniCssExtractPlugin.loader, 'css-loader', 'sass-loader'],
+      // },
       {
         test: /\.(png|jpe?g|gif|eot|woff2|woff|ttf|svg)$/i,
         use: 'file-loader',
-      }
+      },
     ],
   },
   resolve: {
     modules: ['node_modules'],
     extensions: ['.js', '.jsx', '.scss', '.css'],
   },
+  // doesn't seem to do much tree shaking? disabling because i don't trust it for now.
+  // optimization: {
+  //   splitChunks: {
+  //     chunks: 'all',
+  //   },
+  // },
   output: {
     filename: "[name].js",
     sourceMapFilename: "[name].js.map",
     path: path.resolve(__dirname, "../../app/assets/builds"),
   },
   plugins: [
-    new webpack.optimize.LimitChunkCountPlugin({
-      maxChunks: 1
-    }),
-    new FixStyleOnlyEntriesPlugin(),
-    new MiniCssExtractPlugin(),
+    new webpack.ProgressPlugin(), // show progress via --progress
+    // new MiniCssExtractPlugin(), //not using css in webpack - using rails asset pipeline instead
+    // new BundleAnalyzerPlugin() //uncomment to analyze bundle size
   ]
 }
