+++ package.json
@@ -30,10 +30,12 @@
     "node": "^14.x",
     "yarn": "^1.6.0"
   },
-  "devDependencies": {},
+  "devDependencies": {
+    "webpack-bundle-analyzer": "^4.5.0"
+  },
   "license": "MIT",
   "scripts": {
-    "build": "webpack --config ./config/webpack/webpack.config.js",
-    "watch": "webpack --watch --config ./config/webpack/webpack.config.js --mode=development"
+    "build":   "webpack --config ./config/webpack/webpack.config.js --progress --mode=production",
+    "watch": "webpack --watch --config ./config/webpack/webpack.config.js --mode=development --progress"
   }
 }
